Finished exercise 4.23

+++ part4/blog-list/tests/blog_api.test.js
@@ -3,13 +3,19 @@ const supertest = require('supertest')
 const app = require('../app')
 const api = supertest(app)
 const helper = require('./test_helper')
-const Blog = require('../models/blogschema')
-// npm test -- tests/blog_api.test.js
+const Blog = require('../models/blogSchema')
+const User = require('../models/userSchema')
+const jwt = require('jsonwebtoken');
+const bcrypt = require('bcrypt');
+const config = require('../utils/config')
+
+
+// npm test -- -t 'addition of a new blog'
+// npm test -- blog_api.test.js
 
 beforeEach(async () => {
     await Blog.deleteMany({})
     await Blog.insertMany(helper.initialBlogs)
-
 })
 
 describe('when there is initially some blogs saved', () => {
@@ -26,6 +32,15 @@ describe('when there is initially some blogs saved', () => {
 })
 
 describe('addition of a new blog', () => {
+    let token = null
+    beforeEach(async () => {
+        await User.deleteMany({});
+        const passwordHash = await bcrypt.hash("12345", 10);
+        const user = await new User({ username: "name", passwordHash }).save();
+
+        const userForToken = { username: "name", id: user.id };
+        return (token = jwt.sign(userForToken, config.SECRET));
+    })
     test('succeeds with valid data', async () => {
         const newBlog = {
             title: "TDD harms architecture",
@@ -33,7 +48,14 @@ describe('addition of a new blog', () => {
             url: "http://blog.cleancoder.com/uncle-bob/2017/03/03/TDD-Harms-Architecture.html",
             likes: 0,
         }
-        await api.post('/api/blogs').send(newBlog).expect(201).expect('Content-Type', /application\/json/)
+        await api
+            .post('/api/blogs')
+            .set('Authorization', `Bearer ${token}`)
+            .send(newBlog)
+            .expect(201)
+            .expect('Content-Type', /application\/json/)
+        
+        
         const blogsAtEnd = await helper.blogsInDB()
         const contents = blogsAtEnd.map(blog => blog.title)
         expect(blogsAtEnd).toHaveLength(helper.initialBlogs.length + 1)
@@ -47,7 +69,7 @@ describe('addition of a new blog', () => {
             author: "Robert C. Martin",
             url: "http://blog.cleancoder.com/uncle-bob/2017/03/03/TDD-Harms-Architecture.html",
         }
-        await api.post('/api/blogs').send(newBlog).expect(201).expect('Content-Type', /application\/json/)
+        await api.post('/api/blogs').set('Authorization', `Bearer ${token}`).send(newBlog).expect(201).expect('Content-Type', /application\/json/)
         const blogsAtEnd = await helper.blogsInDB()
         blogsAtEnd.forEach(blog => {
             expect(blog.likes).toBeDefined()
@@ -59,22 +81,48 @@ describe('addition of a new blog', () => {
             title: "TDD harms architecture",
             author: "Robert C. Martin",
         }
-    
+
         const noAuthorBlog = {
             title: "TDD harms architecture",
             url: "http://blog.cleancoder.com/uncle-bob/2017/03/03/TDD-Harms-Architecture.html",
         }
-        await api.post('/api/blogs').send(noUrlBlog).expect(400)
-        await api.post('/api/blogs').send(noAuthorBlog).expect(400)
+        await api.post('/api/blogs').set('Authorization', `Bearer ${token}`).send(noUrlBlog).expect(400)
+        await api.post('/api/blogs').set('Authorization', `Bearer ${token}`).send(noAuthorBlog).expect(400)
+    })
+
+    test('fails when no token is provided', async () => {
+        const testblog = {
+            title: "TDD harms architecture",
+            author: "Robert C. Martin",
+            url: "http://blog.cleancoder.com/uncle-bob/2017/03/03/TDD-Harms-Architecture.html"
+        }
+        await api.post('/api/blogs').set('Authorization', `Bearer `).send(testblog).expect(401)
     })
 })
 
 
 describe('deletion of a new blog', () => {
+    let token = null
+    beforeEach(async () => {
+        await Blog.deleteMany({});
+        await User.deleteMany({});
+        const passwordHash = await bcrypt.hash("12345", 10);
+        const user = await new User({ username: "name", passwordHash }).save();
+        const userForToken = { username: "name", id: user.id };
+        token = jwt.sign(userForToken, process.env.SECRET)
+        const testBlog = {
+            title: "TDD harms architecture",
+            author: "Robert C. Martin",
+            url: "http://blog.cleancoder.com/uncle-bob/2017/03/03/TDD-Harms-Architecture.html",
+        }
+        await api.post('/api/blogs').set('Authorization', `Bearer ${token}`).send(testBlog)
+
+
+    })
     test('succeeds with status code 204 if the id is valid', async () => {
-        const blogsAtStart = await helper.blogsInDB()
+        const blogsAtStart = await Blog.find({}).populate('user')
         const blog = blogsAtStart[0]
-        await api.delete(`/api/blogs/${blog.id}`).expect(204)
+        await api.delete(`/api/blogs/${blog.id}`).set('Authorization', `Bearer ${token}`).expect(204)
         const blogsAtEnd = await helper.blogsInDB()
         expect(blogsAtEnd).toHaveLength(blogsAtStart.length - 1)
     })
@@ -84,7 +132,7 @@ describe('updating a blog with a new number of likes', () => {
         const blogsAtStart = await helper.blogsInDB()
         const blog = blogsAtStart[0]
         const newBlog = {
-            ...blog, 
+            ...blog,
             likes: blog.likes + 10
         }
         const updatedBlog = await api.put(`/api/blogs/${blog.id}`).send(newBlog).expect(200)
